moved linter functions to its own file
